UIMenuを拡張して書き直した
